Rough implementation of fetching the catalog index.
I tried having an action on the store to call the CTS GetCapabilities
request to populate the suggested documents, but I couldn't get
that to work. Instead treat it as local data and populate that
asynchronously at mount time. This results is a lot of repeated
loads whenever the widget is shown or expanded.